+++ include/seqan3/alignment/pairwise/alignment_algorithm.hpp
@@ -31,7 +31,7 @@
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/simd/simd.hpp>
 #include <seqan3/utility/simd/simd_traits.hpp>
-#include <seqan3/core/simd/view_to_simd.hpp>
+#include <seqan3/utility/simd/views/to_simd.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/range/views/drop.hpp>
 #include <seqan3/range/views/get.hpp>
