--- test/performance/core/simd/CMakeLists.txt
@@ -1,1 +1,0 @@
-seqan3_benchmark(view_to_simd_chunk_benchmark.cpp)
