+++ test/performance/utility/simd/views/CMakeLists.txt
@@ -1,1 +1,2 @@
 seqan3_benchmark(iota_simd_benchmark.cpp)
+seqan3_benchmark(to_simd_benchmark.cpp)
