+++ test/snippet/utility/simd/views/to_simd.cpp
