+++ test/unit/utility/simd/views/CMakeLists.txt
@@ -1,1 +1,2 @@
 seqan3_test (iota_simd_test.cpp)
+seqan3_test (to_simd_test.cpp)
